requirements.txt
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.2.0
dj-database-url==2.0.0
python-decouple==3.8
psycopg2==2.9.6
drf-yasg==1.21.6
django-ckeditor==6.7.0